refactor: extract TradingView URL logic into reusable hook
- Create useTradingViewUrl hook to abstract repeated URL construction logic
- Refactor TradingViewV2 and TradingViewPerpsV2 to use the new hook
- Eliminate duplicate code for dev settings URL determination and common parameters
- Maintain existing functionality while improving code maintainability
