+++ packages/kit/src/components/TradingView/TradingViewPerpsV2/TradingViewPerpsV2.tsx
@@ -1,18 +1,12 @@
-import { memo, useCallback, useEffect, useMemo, useRef } from 'react';
-
-import { useCalendars } from 'expo-localization';
+import { memo, useCallback, useEffect, useRef } from 'react';
 
 import { Stack, useOrientation } from '@onekeyhq/components';
 import type { IStackStyle } from '@onekeyhq/components';
-import { useDevSettingsPersistAtom } from '@onekeyhq/kit-bg/src/states/jotai/atoms/devSettings';
-import { TRADING_VIEW_URL } from '@onekeyhq/shared/src/config/appConfig';
 import platformEnv from '@onekeyhq/shared/src/platformEnv';
 import type { IHex } from '@onekeyhq/shared/types/hyperliquid/sdk';
 
-import { useLocaleVariant } from '../../../hooks/useLocaleVariant';
-import { useThemeVariant } from '../../../hooks/useThemeVariant';
 import WebView from '../../WebView';
-import { getTradingViewTimezone } from '../utils/tradingViewTimezone';
+import { useTradingViewUrl } from '../hooks';
 
 import { useTradeUpdates } from './hooks';
 import { usePerpsMessageHandler } from './messageHandlers';
@@ -103,33 +97,18 @@ export function TradingViewPerpsV2(
   const isLandscape = useOrientation();
   const isIPadPortrait = platformEnv.isNativeIOSPad && !isLandscape;
   const webRef = useRef<IWebViewRef | null>(null);
-  const calendars = useCalendars();
-  const systemLocale = useLocaleVariant();
-  const theme = useThemeVariant();
-  const [devSettings] = useDevSettingsPersistAtom();
 
   const { symbol, userAddress, onLoadEnd, onTradeUpdate, tradingViewUrl } =
     props;
 
   // Optimization: Static URL with only initialization params to avoid WebView reload
-  const staticTradingViewUrl = useMemo(() => {
-    const baseUrl =
-      tradingViewUrl ||
-      (devSettings.enabled && devSettings.settings?.useLocalTradingViewUrl
-        ? 'http://localhost:5173/'
-        : TRADING_VIEW_URL);
-
-    const url = new URL(baseUrl);
-    url.searchParams.set('timezone', getTradingViewTimezone(calendars));
-    url.searchParams.set('locale', systemLocale);
-    url.searchParams.set('platform', platformEnv.appPlatform ?? 'web');
-    url.searchParams.set('theme', theme);
-    url.searchParams.set('symbol', symbol);
-    url.searchParams.set('type', 'perps');
-
-    return url.toString();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [tradingViewUrl, devSettings, calendars, systemLocale, theme]);
+  const { finalUrl: staticTradingViewUrl } = useTradingViewUrl({
+    tradingViewUrl,
+    additionalParams: {
+      symbol,
+      type: 'perps',
+    },
+  });
 
   // Optimization: Dynamic symbol parameter sync mechanism
   useSymbolSync({
