+++ packages/kit/src/components/TradingView/TradingViewV2/TradingViewV2.tsx
@@ -1,17 +1,11 @@
-import { useCallback, useMemo, useRef } from 'react';
-
-import { useCalendars } from 'expo-localization';
+import { useCallback, useRef } from 'react';
 
 import { Stack, useOrientation } from '@onekeyhq/components';
 import type { IStackStyle } from '@onekeyhq/components';
-import { useDevSettingsPersistAtom } from '@onekeyhq/kit-bg/src/states/jotai/atoms/devSettings';
-import { TRADING_VIEW_URL } from '@onekeyhq/shared/src/config/appConfig';
 import platformEnv from '@onekeyhq/shared/src/platformEnv';
 
-import { useLocaleVariant } from '../../../hooks/useLocaleVariant';
-import { useThemeVariant } from '../../../hooks/useThemeVariant';
 import WebView from '../../WebView';
-import { getTradingViewTimezone } from '../utils/tradingViewTimezone';
+import { useTradingViewUrl } from '../hooks';
 
 import {
   useAutoKLineUpdate,
@@ -47,10 +41,6 @@ export function TradingViewV2(props: ITradingViewV2Props & WebViewProps) {
   const isLandscape = useOrientation();
   const isIPadPortrait = platformEnv.isNativeIOSPad && !isLandscape;
   const webRef = useRef<IWebViewRef | null>(null);
-  const calendars = useCalendars();
-  const systemLocale = useLocaleVariant();
-  const theme = useThemeVariant();
-  const [devSettings] = useDevSettingsPersistAtom();
 
   const {
     mode,
@@ -71,45 +61,15 @@ export function TradingViewV2(props: ITradingViewV2Props & WebViewProps) {
     onPanesCountChange,
   });
 
-  // Determine the URL to use based on dev settings
-  const finalTradingViewUrl = useMemo(() => {
-    if (tradingViewUrl) {
-      return tradingViewUrl;
-    }
-
-    return devSettings.enabled && devSettings.settings?.useLocalTradingViewUrl
-      ? 'http://localhost:5173/'
-      : TRADING_VIEW_URL;
-  }, [
+  const { finalUrl: tradingViewUrlWithParams } = useTradingViewUrl({
     tradingViewUrl,
-    devSettings.enabled,
-    devSettings.settings?.useLocalTradingViewUrl,
-  ]);
-
-  const tradingViewUrlWithParams = useMemo(() => {
-    const timezone = getTradingViewTimezone(calendars);
-    const locale = systemLocale;
-
-    const url = new URL(finalTradingViewUrl);
-    url.searchParams.set('timezone', timezone);
-    url.searchParams.set('locale', locale);
-    url.searchParams.set('platform', platformEnv.appPlatform ?? 'web');
-    url.searchParams.set('theme', theme);
-    url.searchParams.set('symbol', symbol);
-    url.searchParams.set('decimal', decimal?.toString());
-    url.searchParams.set('networkId', networkId);
-    url.searchParams.set('address', tokenAddress);
-    return url.toString();
-  }, [
-    finalTradingViewUrl,
-    calendars,
-    systemLocale,
-    theme,
-    symbol,
-    decimal,
-    networkId,
-    tokenAddress,
-  ]);
+    additionalParams: {
+      symbol,
+      decimal: decimal?.toString(),
+      networkId,
+      address: tokenAddress,
+    },
+  });
 
   useAutoKLineUpdate({
     tokenAddress,
