+++ packages/kit/src/components/TradingView/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useTradingViewUrl } from './useTradingViewUrl';
